test for mysql function and procedures

+++ server/models/procedures/migration-process.sql
@@ -42,13 +42,29 @@ DELIMITER ;
 
 -- From  https://stackoverflow.com/questions/2480148/how-can-i-employ-if-exists-for-creating-or-dropping-an-index-in-mysql
 -- This procedure try to drop a table index if it exists
+
+DELIMITER $$
+DROP FUNCTION IF EXISTS index_exists;
+
+CREATE FUNCTION index_exists(
+  theTable VARCHAR(64),
+  theIndexName VARCHAR(64)
+)
+  RETURNS BOOLEAN
+  READS SQL DATA
+  BEGIN
+    RETURN 0 < (SELECT COUNT(*) AS exist FROM information_schema.statistics WHERE TABLE_SCHEMA = DATABASE() and table_name =
+theTable AND index_name = theIndexName);
+  END $$
+DELIMITER ;
+
+
 DELIMITER $$
 
 DROP PROCEDURE IF EXISTS drop_index_if_exists $$
 CREATE PROCEDURE drop_index_if_exists(in theTable varchar(128), in theIndexName varchar(128) )
 BEGIN
- IF((SELECT COUNT(*) AS index_exists FROM information_schema.statistics WHERE TABLE_SCHEMA = DATABASE() and table_name =
-theTable AND index_name = theIndexName) > 0) THEN
+ IF(index_exists (theTable, theIndexName)) THEN
    SET @s = CONCAT('DROP INDEX ' , theIndexName , ' ON ' , theTable);
    PREPARE stmt FROM @s;
    EXECUTE stmt;
@@ -56,3 +72,49 @@ theTable AND index_name = theIndexName) > 0) THEN
 END $$
 
 DELIMITER ;
+
+
+-- 
+
+DELIMITER $$
+DROP FUNCTION IF EXISTS Constraint_exists;
+
+CREATE FUNCTION Constraint_exists(
+  theTable VARCHAR(64),
+  theConstraintName VARCHAR(64)
+)
+  RETURNS BOOLEAN
+  READS SQL DATA
+  BEGIN
+    RETURN 0 < (
+		 SELECT COUNT(*) AS nbr
+		 FROM
+	    INFORMATION_SCHEMA.REFERENTIAL_CONSTRAINTS
+	   WHERE CONSTRAINT_SCHEMA = DATABASE()
+	   AND TABLE_NAME= theTable
+		AND  CONSTRAINT_NAME = theConstraintName
+	 );
+  END $$
+DELIMITER ;
+
+
+
+
+DELIMITER $$
+
+-- this procedure will be used for "ALTER TABLE table_name DROP FOREIGN KEY constraint_name";
+-- example : CALL drop_foreign_key('table_name', 'constraint_name');
+
+DROP PROCEDURE IF EXISTS drop_foreign_key $$
+CREATE PROCEDURE drop_foreign_key(in theTable varchar(128), in theConstraintName varchar(128) )
+BEGIN
+ IF(Constraint_exists(theTable, theConstraintName) > 0) THEN
+ 
+   SET @s = CONCAT(' ALTER TABLE ' , theTable , ' DROP FOREIGN KEY  ' , theConstraintName);
+   PREPARE stmt FROM @s;
+   EXECUTE stmt;
+ END IF;
+END $$
+
+DELIMITER ;
+
